+++ src/mat/4C_mat_material_factory.cpp
@@ -52,10 +52,6 @@
 #include "4C_mat_elast_isovarga.hpp"
 #include "4C_mat_elast_isoyeoh.hpp"
 #include "4C_mat_elast_remodelfiber.hpp"
-#include "4C_mat_elast_visco_coupmyocard.hpp"
-#include "4C_mat_elast_visco_fsls.hpp"
-#include "4C_mat_elast_visco_generalizedmaxwell.hpp"
-#include "4C_mat_elast_visco_isoratedep.hpp"
 #include "4C_mat_elast_vologden.hpp"
 #include "4C_mat_elast_volpenalty.hpp"
 #include "4C_mat_elast_volpow.hpp"
@@ -145,6 +141,10 @@
 #include "4C_mat_thermoplasticlinelast.hpp"
 #include "4C_mat_thermostvenantkirchhoff.hpp"
 #include "4C_mat_viscoanisotropic.hpp"
+#include "4C_mat_viscoelast_coupmyocard.hpp"
+#include "4C_mat_viscoelast_fsls.hpp"
+#include "4C_mat_viscoelast_generalizedmaxwell.hpp"
+#include "4C_mat_viscoelast_isoratedep.hpp"
 #include "4C_mat_viscoelasthyper.hpp"
 #include "4C_mat_visconeohooke.hpp"
 #include "4C_mat_viscoplastic_no_yield_surface.hpp"
