+++ src/mat/elast/4C_mat_elast_summand.cpp
@@ -42,15 +42,15 @@
 #include "4C_mat_elast_isovarga.hpp"
 #include "4C_mat_elast_isoyeoh.hpp"
 #include "4C_mat_elast_remodelfiber.hpp"
-#include "4C_mat_elast_visco_coupmyocard.hpp"
-#include "4C_mat_elast_visco_fsls.hpp"
-#include "4C_mat_elast_visco_generalizedmaxwell.hpp"
-#include "4C_mat_elast_visco_isoratedep.hpp"
 #include "4C_mat_elast_vologden.hpp"
 #include "4C_mat_elast_volpenalty.hpp"
 #include "4C_mat_elast_volpow.hpp"
 #include "4C_mat_elast_volsussmanbathe.hpp"
 #include "4C_mat_par_bundle.hpp"
+#include "4C_mat_viscoelast_coupmyocard.hpp"
+#include "4C_mat_viscoelast_fsls.hpp"
+#include "4C_mat_viscoelast_generalizedmaxwell.hpp"
+#include "4C_mat_viscoelast_isoratedep.hpp"
 #include "4C_solid_ele_fibers.hpp"
 #include "4C_utils_enum.hpp"
