+++ src/mat/viscoelast/4C_mat_viscoelast_isoratedep.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_mat_elast_visco_isoratedep.hpp"
+#include "4C_mat_viscoelast_isoratedep.hpp"
 
 #include "4C_material_parameter_base.hpp"
